Create registerCommands.ts
